test: refactor directory structure
break tests up into new directory structure
to differentiate between integration tests, unit
tests and end to end tests. hopefully this will make
it more obvious to contributors what parts of the
codebase are covered and make it easier to contribute
