+++ package.json
@@ -66,8 +66,9 @@
     "lint": "eslint $(cat .eslintfiles) || exit 0",
     "lint-ci": "eslint $(cat .eslintfiles)",
     "lint-file": "eslint",
-    "test": "bmocha --reporter spec test/*.js",
-    "test-browser": "NODE_BACKEND=js bmocha --reporter spec test/*.js",
+    "test": "bmocha --reporter spec test/**/*.js",
+    "test-integration": "bmocha --reporter spec test/integration/*.js",
+    "test-browser": "NODE_BACKEND=js bmocha --reporter spec test/**/*.js test/*.js",
     "test-file": "bmocha --reporter spec",
     "test-file-browser": "NODE_BACKEND=js bmocha --reporter spec",
     "test-ci": "istanbul cover --report lcovonly node_modules/.bin/bmocha -- --reporter spec test/*.js"
