+++ test/README.md
@@ -1,0 +1,18 @@
+# Testing
+
+## Usage
+
+```bash
+$ npm run test
+```
+
+```bash
+$ npm run test-integration
+```
+
+## Overview
+
+
+## Contributing
+
+
