+++ test/coins/coins-test.js
@@ -1,0 +1,17 @@
+/* eslint-env mocha */
+/* eslint prefer-arrow-callback: "off" */
+/* eslint no-unused-vars: "off" */
+
+'use strict';
+
+const bio = require('bufio');
+const assert = require('../util/assert');
+const Output = require('../../lib/primitives/output');
+const Input = require('../../lib/primitives/input');
+const Outpoint = require('../../lib/primitives/outpoint');
+const CoinView = require('../../lib/coins/coinview');
+const CoinEntry = require('../../lib/coins/coinentry');
+const common = require('../util/common');
+
+describe('Coins', function() {
+});
