+++ test/integration/wallet-test.js
@@ -3,22 +3,23 @@
 
 'use strict';
 
-const assert = require('./util/assert');
-const consensus = require('../lib/protocol/consensus');
-const Network = require('../lib/protocol/network');
-const util = require('../lib/utils/util');
+const assert = require('../util/assert');
+const consensus = require('../../lib/protocol/consensus');
+const Network = require('../../lib/protocol/network');
+const util = require('../../lib/utils/util');
 const blake2b = require('bcrypto/lib/blake2b');
 const random = require('bcrypto/lib/random');
-const WalletDB = require('../lib/wallet/walletdb');
-const WorkerPool = require('../lib/workers/workerpool');
-const Address = require('../lib/primitives/address');
-const MTX = require('../lib/primitives/mtx');
-const Coin = require('../lib/primitives/coin');
-const KeyRing = require('../lib/primitives/keyring');
-const Input = require('../lib/primitives/input');
-const Outpoint = require('../lib/primitives/outpoint');
-const Script = require('../lib/script/script');
-const PrivateKey = require('../lib/hd/private.js');
+const WalletDB = require('../../lib/wallet/walletdb');
+const WorkerPool = require('../../lib/workers/workerpool');
+const Address = require('../../lib/primitives/address');
+const MTX = require('../../lib/primitives/mtx');
+const Coin = require('../../lib/primitives/coin');
+const KeyRing = require('../../lib/primitives/keyring');
+const Input = require('../../lib/primitives/input');
+const Outpoint = require('../../lib/primitives/outpoint');
+const Script = require('../../lib/script/script');
+const HD = require('../../lib/hd');
+const PrivateKey = require('../../lib/hd/private.js');
 
 const KEY1 = 'xprv9s21ZrQH143K3Aj6xQBymM31Zb4BVc7wxqfUhMZrzewdDVCt'
   + 'qUP9iWfcHgJofs25xbaUpCps9GDXj83NiWvQCAkWQhVj5J4CorfnpKX94AZ';
