+++ test/primitives/address-test.js
@@ -3,8 +3,8 @@
 
 'use strict';
 
-const assert = require('./util/assert');
-const Address = require('../lib/primitives/address');
+const assert = require('../util/assert');
+const Address = require('../../lib/primitives/address');
 
 describe('Address', function() {
   it('should match mainnet p2pkh address', () => {
