+++ test/primitives/input-test.js
@@ -5,9 +5,9 @@
 'use strict';
 
 const bio = require('bufio');
-const Input = require('../lib/primitives/input');
-const assert = require('./util/assert');
-const common = require('./util/common');
+const Input = require('../../lib/primitives/input');
+const assert = require('../util/assert');
+const common = require('../util/common');
 
 describe('Input', function() {
 });
